requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
openpyxl
pandas
pydantic-settings
python-dotenv
ipykernel
matplotlib